Homework 6 spring orm

+++ src/main/liquibase/sql/01.create-books-table.sql
@@ -1,8 +1,8 @@
 create table books(
-id serial primary key,
+id bigserial primary key,
 name_book varchar(64),
-genre_id int references genres(id),
-language_id int references languages(id),
+genre_id int8 references genres(id),
+language_id int8 references languages(id),
 number_of_page int,
 year_of_publishing int,
 number_of_copies int,
