+++ src/main/liquibase/sql/06.create-book_publisher-table.sql
@@ -1,4 +1,4 @@
 create table book_publisher(
-publisher_id int references publishers(id),
-book_id int references  books(id)
+publisher_id int8 references publishers(id),
+book_id int8 references  books(id)
 );
