+++ src/main/liquibase/sql/10.create-user_role-table.sql
@@ -1,4 +1,4 @@
 create table user_role(
-role_id int references roles(id),
-user_id int references users(id)
+role_id int8 references roles(id),
+user_id int8 references users(id)
 );
